Refactor controllers: Move newGame() and enterGame() methods from LoginController.java to NewgameController.java
